supporting markAsDelete in script.

+++ src/main/java/com/facilio/workflowv2/modulefunctions/FacilioModuleFunctionImpl.java
@@ -339,7 +339,12 @@ public class FacilioModuleFunctionImpl implements FacilioModuleFunction {
 		
 		delete.andCriteria(criteria);
 		
-		delete.delete();
+		if (module.isTrashEnabled()) {
+            delete.markAsDelete();
+        }
+		else {
+			delete.delete();
+		}
 	}
 	
 	@Override
